requirements.txt
## These requirements were autogenerated by pipenv# To regenerate from the project's Pipfile, run:##    pipenv lock --requirements#-i https://pypi.org/simplebotocore==1.19.11certifi==2020.6.20click==7.1.2configparser==5.0.1etcd3==0.12.0grpc-interceptor==0.12.0grpcio-tools==1.27.2grpcio==1.27.2jmespath==0.10.0minio==6.0.0pathtools==0.1.2protobuf==3.13.0python-dateutil==2.8.1python-decouple==3.3pytz==2020.4six==1.15.0tenacity==6.2.0urllib3==1.25.11watchdog==0.10.4